Adapted to new operator structure (as far as possible). Updated to operator-rs 0.7.0. Removed packaging and operator crate. Updated docs. Fixed port on 3030 (no CLI anymore).
