+++ docs/modules/ROOT/pages/installation.adoc
@@ -2,82 +2,29 @@
 
 There are three ways to run the Regorule Operator:
 
-1. Build from source. This is documented in xref:building.adoc[Building the Operator].
+1. Helm managed Docker container deployment on Kubernetes
 
-2. Operating System packages (dep & RPM)
+2. As a Docker container
 
-3. As a Docker container
+3. Build from source.
 
-== Operating System Packages
+== Helm
 
-=== Debian
-
-We do provide `deb` packages for Debian 10:
-
-|===
-| Version | Repository
-
-| Development
-| `https://repo.stackable.tech/repository/deb-dev/ buster main`
-
-| Stable
-| `https://repo.stackable.tech/repository/deb-release/ buster main`
-|===
-
-To add these repositories run a command like this:
-
+Helm allows you to download and deploy Stackable operators on Kubernetes and is by far the easiest
+installation method. First ensure that you have installed the Stackable Operators Helm repository:
 [source,bash]
 ----
-echo "deb https://repo.stackable.tech/repository/deb-release/ buster main" | sudo tee -a /etc/apt/sources.list.d/stackable.list
+$ helm repo add stackable https://repo.stackable.tech/repository/helm-stable/
 ----
 
-After you added the repository run:
-
+Then install the Stackable Regorule Operator
 [source,bash]
 ----
-sudo apt-get update
-sudo apt-get install stackable-regorule-operator
+$ helm install regorule-operator stackable/regorule-operator
 ----
 
-=== RHEL/CentOS
-
-We do provide `RPM` packages for RHEL 7 & 8.
-
-|===
-| Version | Repository
-
-| Development RHEL 7
-| `https://repo.stackable.tech/repository/rpm-dev/el7`
-
-| Stable RHEL 7
-| `https://repo.stackable.tech/repository/rpm-release/el7`
-
-| Development RHEL 8
-| `https://repo.stackable.tech/repository/rpm-dev/el8`
-
-| Stable RHEL 8
-| `https://repo.stackable.tech/repository/rpm-release/el8`
-
-|===
-
-Add this to `/etc/yum.repos.d/stackable.repo` and make sure to pick the corrrect repository:
-
-[source]
-----
-[stackable]
-name=Stackable
-baseurl=https://repo.stackable.tech/repository/rpm-release/el8
-repo_gpgcheck=0
-enabled=1
-gpgcheck=0
-----
-
-Install the Operator:
-
-[source,bash]
-----
-sudo yum install stackable-regorule-operator
-----
+Helm will deploy the operator in a Kubernetes container and apply the CRDs for the Regorule
+service. You are now ready to deploy the Stackable Regorule Operator in Kubernetes.
 
 == Docker
 
@@ -88,8 +35,32 @@ This Operator is published as a Docker image:
 docker.stackable.tech/stackable/regorule-operator
 ----
 
+When installing manually with Docker you will need to install the Stackable CRDs for the Regorule Operator
+in your Kubernetes environment. These are available on the
+https://github.com/stackabletech/regorule-operator/tree/main/deploy/crd[Stackable GitHub repository]
+for this operator.
+[source]
+----
+$ kubectl apply -f regorule.crd.yaml
+----
+
 To run it straight from Docker you can use this command:
 [source,bash]
 ----
-docker run -e REGORULE_OPERATOR_LOG=trace -e KUBECONFIG=/root/.kube/k3s.yaml --mount type=bind,source="$HOME/.kube/k3s.yaml",target="/root/.kube/k3s.yaml" -it docker.stackable.tech/stackable/regorule-operator:latest
+docker run \
+    --name regorule-operator \
+    --network host \
+    --env KUBECONFIG=/home/stackable/.kube/config \
+    --mount type=bind,source="$HOME/.kube/config",target="/home/stackable/.kube/config" \
+    docker.stackable.tech/stackable/regorule-operator:latest
 ----
+
+== Building the operator from source
+
+This operator is written in Rust and is developed against the latest stable Rust release (1.57 at
+the time of writing).
+
+[source]
+----
+cargo run -- crd | kubectl apply -f -
+cargo run -- run
