--- packaging/debian/stackable-regorule-operator.service
@@ -1,13 +1,0 @@
-[Unit]
-Description=Stackable Operator for RegoRules
-Before=
-After=network.target
-[Service]
-User=root
-ExecStart=/opt/stackable/regorule-operator/stackable-regorule-operator
-Restart=on-abort
-StandardOutput=journal
-StandardError=journal
-Environment="REGORULE_OPERATOR_LOG=info"
-[Install]
-WantedBy=multi-user.target
