+++ rust/crd/Cargo.toml
@@ -1,21 +1,14 @@
 [package]
-authors = ["Lars Francke <lars.francke@stackable.de>"]
+authors = ["Stackable GmbH <info@stackable.de>"]
 description = "Contains the RegoRule CRD structs and utilities"
-edition = "2018"
+edition = "2021"
 license = "OSL-3.0"
 name = "stackable-regorule-crd"
 repository = "https://github.com/stackabletech/regorule-operator"
 version = "0.5.0-nightly"
 
 [dependencies]
-stackable-operator = { git = "https://github.com/stackabletech/operator-rs.git", tag = "0.4.0" }
+stackable-operator = { git = "https://github.com/stackabletech/operator-rs.git", tag = "0.7.0" }
 
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-
-[dev-dependencies]
-
-[features]
-default = ["native-tls"]
-native-tls = ["stackable-operator/native-tls"]
-rustls-tls = ["stackable-operator/rustls-tls"]
