--- rust/operator/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-authors = ["Lars Francke <lars.francke@stackable.de>"]
-description = "Stackable Operator for OpenPolicyAgent Rego rules - The Operator Library"
-edition = "2018"
-license = "OSL-3.0"
-name = "stackable-regorule-operator"
-repository = "https://github.com/stackabletech/regorule-operator"
-version = "0.5.0-nightly"
-
-[dependencies]
-stackable-regorule-crd = { path = "../crd" }
-stackable-operator = { git = "https://github.com/stackabletech/operator-rs.git", tag = "0.4.0" }
-
-async-trait = "0.1"
-flate2 = "1.0"
-futures = "0.3"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-tar = "0.4"
-thiserror = "1.0"
-tokio = "1.15"
-tracing = "0.1"
-warp = "0.3"
-
-[dev-dependencies]
-rstest = "0.11"
